+++ src/collective/collectionfilter/portlets/collectionsearch.py
@@ -4,22 +4,13 @@ from ..baseviews import BaseSearchView
 from ..interfaces import ICollectionSearchSchema
 from plone.app.portlets.portlets import base
 from plone.portlets.interfaces import IPortletDataProvider
-from ..utils import get_top_request
-from Products.CMFPlone.utils import getFSVersionTuple
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from zope.component import queryUtility
 from zope.interface import implementer
 
-
-PLONE5 = getFSVersionTuple()[0] >= 5
-
-if PLONE5:
-    base_AddForm = base.AddForm
-    base_EditForm = base.EditForm
-else:
-    from plone.app.portlets.browser.z3cformhelper import AddForm as base_AddForm  # noqa
-    from plone.app.portlets.browser.z3cformhelper import EditForm as base_EditForm  # noqa
-    from z3c.form import field
+try:
+    from Products.CMFPlone.utils import get_top_request
+except ImportError:
+    from collective.collectionfilter.utils import get_top_request
 
 
 class ICollectionSearchPortlet(ICollectionSearchSchema, IPortletDataProvider):
@@ -76,12 +67,9 @@ class Renderer(base.Renderer, BaseSearchView):
         return reload_url
 
 
-class AddForm(base_AddForm):
-    if PLONE5:
-        schema = ICollectionSearchPortlet
-    else:
-        fields = field.Fields(ICollectionSearchPortlet)
+class AddForm(base.AddForm):
 
+    schema = ICollectionSearchPortlet
     label = _(u"Add Collection Search Portlet")
     description = _(
         u"This portlet allows fulltext search in collection results."
@@ -91,12 +79,9 @@ class AddForm(base_AddForm):
         return Assignment(**data)
 
 
-class EditForm(base_EditForm):
-    if PLONE5:
-        schema = ICollectionSearchPortlet
-    else:
-        fields = field.Fields(ICollectionSearchPortlet)
+class EditForm(base.EditForm):
 
+    schema = ICollectionSearchPortlet
     label = _(u"Edit Collection Search Portlet")
     description = _(
         u"This portlet allows fulltext search in collection results."